Refactor project structure, remove deprecated files, and update modules

+++ src/modules/grade/grade.service.js
@@ -1,5 +1,8 @@
 const Grade = require("./grade.model");
 const ApiError = require("../../utils/ApiErrors");
+const Group = require("../group/group.model");
+const Teacher = require('../teacher/teacher.model');
+const Student = require('../student/student.model');
 
 // Create Grade
 exports.createGrade = async (data) => {
@@ -15,10 +18,31 @@ exports.createGrade = async (data) => {
 };
 
 // Get All Grades
-exports.getAllGrades = async () => {
-  const grades = await Grade.find();
+exports.getAllGrades = async (user) => {
+  let filter = {};
 
-  return grades;
+  if (!user) {
+    return await Grade.find(filter);
+  }
+
+  if (user.role === "teacher" || user.role === "secretary") {
+    let teacherUserId = user.role === "teacher" ? (user._id || user.id) : user.createdBy;
+    const teacher = await Teacher.findOne({ userId: teacherUserId });
+    
+    if (!teacher) return [];
+
+    const teacherGroups = await Group.find({ teacherId: teacher._id }).select("gradeLevelId");
+    const gradeIds = teacherGroups.map(g => g.gradeLevelId);
+
+    filter = { _id: { $in: gradeIds } };
+  } 
+  else if (user.role === "student") {
+    const student = await Student.findOne({ userId: user._id || user.id });
+    if (!student) return [];
+    filter = { _id: student.grade }; 
+  }
+
+  return await Grade.find(filter);
 };
 
 // Get Grade By Id
@@ -40,7 +64,6 @@ exports.updateGrade = async (id, data) => {
     throw new ApiError("Grade not found", 404);
   }
 
-  // التأكد إن الاسم الجديد مش موجود
   if (data.name) {
     const gradeExists = await Grade.findOne({
       name: data.name,
@@ -66,6 +89,14 @@ exports.deleteGrade = async (id) => {
 
   if (!grade) {
     throw new ApiError("Grade not found", 404);
+  }
+
+  const groupsUsingGrade = await Group.exists({ gradeLevelId: id });
+  if (groupsUsingGrade) {
+    throw new ApiError(
+      "Cannot delete grade: it is still used by one or more groups",
+      400
+    );
   }
 
   await Grade.findByIdAndDelete(id);
